# docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "# Document of TSS-microservices APIs\n\nthere are no TOS at this moment, use at your own risk we take no responsibility",
    "title": "TSS-microservices APIs",
    "contact": {
      "name": "Pham Hai",
      "email": "pthai.it.210501@gmail.com"
    },
    "version": "0.0.1"
  },
  "host": "localhost",
  "basePath": "/api/v1",
  "paths": {
    "/feedback": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Feedback"
        ],
        "summary": "Get much feedback as pagination",
        "operationId": "GetManyFeedBack",
        "parameters": [
          {
            "type": "string",
            "description": "maximum number of feedback need to return",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "description": "offset of feedback need to return",
            "name": "offset",
            "in": "query"
          },
          {
            "type": "string",
            "description": "type of feedback need to return",
            "name": "type",
            "in": "query"
          },
          {
            "type": "string",
            "description": "decide list feedback need to return is bug or not",
            "name": "is_bug",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GetManyResponse"
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Feedback"
        ],
        "summary": "Create a feedback",
        "operationId": "CreateFeedback",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateFeedbackFormRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "created feedback",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "$ref": "#/definitions/FeedbackModel"
                }
              }
            }
          },
          "400": {
            "$ref": "#/responses/ValidationError"
          }
        }
      }
    },
    "/feedback/{id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Feedback"
        ],
        "summary": "Get a feedback by id",
        "operationId": "GetFeedbackById",
        "parameters": [
          {
            "type": "string",
            "description": "id of feedback need to return",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "feedback",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "$ref": "#/definitions/FeedbackModel"
                }
              }
            }
          }
        }
      }
    },
    "/hermes/send-mail": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Hermes"
        ],
        "summary": "Send mail",
        "operationId": "SendMail",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SendMailFormRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "send email",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/responses/ValidationError"
          }
        }
      }
    },
    "/teachers": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Teacher"
        ],
        "summary": "Create a teacher",
        "operationId": "CreateTeacher",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Name",
            "name": "name",
            "in": "query",
            "required": true
          },
          {
            "type": "boolean",
            "x-go-name": "IsFemale",
            "name": "is_female",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "format": "date",
            "x-go-name": "Birth",
            "name": "birth",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "UniversityTeacherDegree",
            "x-nullable": "true",
            "name": "university_teacher_degree",
            "in": "query"
          },
          {
            "type": "boolean",
            "x-go-name": "IsHeadOfDepartment",
            "description": "description: nullable",
            "name": "is_head_of_department",
            "in": "query"
          },
          {
            "type": "boolean",
            "x-go-name": "IsHeadOfFaculty",
            "name": "is_head_of_faculty",
            "in": "query"
          },
          {
            "type": "boolean",
            "x-go-name": "IsActive",
            "name": "is_active",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "DepartmentId",
            "name": "department_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "teacher",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "$ref": "#/definitions/TeacherModel"
                }
              }
            }
          }
        }
      }
    },
    "/teachers/:id": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Teacher"
        ],
        "summary": "Get a teacher-form by id",
        "operationId": "GetTeacherById",
        "parameters": [
          {
            "type": "string",
            "description": "id of teacher need to return",
            "name": "id",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "teacher",
            "schema": {
              "type": "object",
              "properties": {
                "data": {
                  "$ref": "#/definitions/TeacherModel"
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "CreateFeedbackFormRequest": {
      "type": "object",
      "required": [
        "data",
        "type",
        "is_bug",
        "account_id"
      ],
      "properties": {
        "account_id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "AccountId"
        },
        "data": {
          "$ref": "#/definitions/FeedbackData"
        },
        "is_bug": {
          "type": "boolean",
          "x-go-name": "IsBug"
        },
        "type": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Type"
        }
      },
      "x-go-package": "TSS-microservices/core/http/requests/feedback-form"
    },
    "DeletedAt": {
      "$ref": "#/definitions/NullTime"
    },
    "FeedbackData": {
      "type": "object",
      "required": [
        "title",
        "content"
      ],
      "properties": {
        "content": {
          "type": "string",
          "x-go-name": "Content"
        },
        "title": {
          "type": "string",
          "x-go-name": "Title"
        }
      },
      "x-go-package": "TSS-microservices/core/models"
    },
    "FeedbackModel": {
      "type": "object",
      "properties": {
        "account_id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "AccountId"
        },
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "data": {
          "$ref": "#/definitions/FeedbackData"
        },
        "id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "ID"
        },
        "is_bug": {
          "type": "boolean",
          "x-go-name": "IsBug"
        },
        "type": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Type"
        }
      },
      "x-go-name": "Feedback",
      "x-go-package": "TSS-microservices/core/models"
    },
    "GetManyResponseBody": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FeedbackModel"
          },
          "x-go-name": "Data"
        },
        "limit": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Limit"
        },
        "offset": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Offset"
        }
      },
      "x-go-package": "TSS-microservices/core/http/responses/feedback"
    },
    "NullTime": {
      "description": "NullTime implements the [Scanner] interface so\nit can be used as a scan destination, similar to [NullString].",
      "type": "object",
      "title": "NullTime represents a [time.Time] that may be null.",
      "properties": {
        "Time": {
          "type": "string",
          "format": "date-time"
        },
        "Valid": {
          "type": "boolean"
        }
      },
      "x-go-package": "database/sql"
    },
    "ResponseBody": {
      "type": "object",
      "properties": {
        "data": {
          "x-go-name": "Data"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-name": "Body",
      "x-go-package": "TSS-microservices/core/http/responses"
    },
    "SendMailFormRequest": {
      "type": "object",
      "required": [
        "code",
        "recipients"
      ],
      "properties": {
        "code": {
          "type": "string",
          "x-go-name": "Code"
        },
        "placeholder": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Placeholder"
        },
        "recipients": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Recipients",
          "example": [
            "email1@gmal.com",
            "email2@gmail.com"
          ]
        },
        "subject": {
          "type": "string",
          "x-go-name": "Subject",
          "x-note": "All key must be capital the first character"
        }
      },
      "x-go-package": "TSS-microservices/core/http/requests/hermes-forms"
    },
    "TeacherModel": {
      "type": "object",
      "properties": {
        "birth": {
          "type": "string",
          "format": "date",
          "x-go-name": "Birth"
        },
        "deleted_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "DeletedAt"
        },
        "department_id": {
          "type": "string",
          "x-go-name": "DepartmentId",
          "x-property-value": "value"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "is_active": {
          "type": "boolean",
          "x-go-name": "IsActive"
        },
        "is_female": {
          "type": "boolean",
          "x-go-name": "IsFemale"
        },
        "is_head_of_department": {
          "type": "boolean",
          "x-go-name": "IsHeadOfDepartment"
        },
        "is_head_of_faculty": {
          "type": "boolean",
          "x-go-name": "IsHeadOfFaculty"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "university_teacher_degree": {
          "type": "string",
          "x-go-name": "UniversityTeacherDegree",
          "x-nullable": true,
          "example": "university_teacher_degree"
        }
      },
      "x-go-name": "Teacher",
      "x-go-package": "TSS-microservices/core/models"
    }
  },
  "responses": {
    "GetManyResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/GetManyResponseBody"
      }
    },
    "ValidationError": {
      "description": "A ValidationError is an error that is used when the required input fails validation.",
      "schema": {
        "$ref": "#/definitions/ResponseBody"
      }
    }
  },
  "tags": [
    {
      "description": "Operations related to teacher management",
      "name": "Teacher"
    },
    {
      "description": "Operations related to feedback management",
      "name": "Feedback"
    }
  ]
}